feat: refactor project structure into feature-based folders
