--- src/Order.Data/IOrderRepository.cs
@@ -1,14 +1,0 @@
-﻿using Order.Model;
-using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace Order.Data
-{
-    public interface IOrderRepository
-    {
-        Task<IEnumerable<OrderSummary>> GetOrdersAsync();
-        Task<OrderDetail> GetOrderByIdAsync(Guid orderId);
-        Task<IEnumerable<OrderSummary>> GetOrdersByStatusAsync(string statusName);
-    }
-}
