--- src/Order.Data/OrderRepository.cs
@@ -1,111 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Order.Model;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Order.Data
-{
-    public class OrderRepository : IOrderRepository
-    {
-        private readonly OrderContext _orderContext;
-
-        public OrderRepository(OrderContext orderContext)
-        {
-            _orderContext = orderContext;
-        }
-
-        public async Task<IEnumerable<OrderSummary>> GetOrdersAsync()
-        {
-            var orders = await _orderContext.Order
-                .Include(x => x.Items)
-                .Include(x => x.Status)
-                .Select(x => new OrderSummary
-                {
-                    Id = new Guid(x.Id),
-                    ResellerId = new Guid(x.ResellerId),
-                    CustomerId = new Guid(x.CustomerId),
-                    StatusId = new Guid(x.StatusId),
-                    StatusName = x.Status.Name,
-                    ItemCount = x.Items.Count,
-                    TotalCost = x.Items.Sum(i => i.Quantity * i.Product.UnitCost).Value,
-                    TotalPrice = x.Items.Sum(i => i.Quantity * i.Product.UnitPrice).Value,
-                    CreatedDate = x.CreatedDate
-                })
-                .OrderByDescending(x => x.CreatedDate)
-                .ToListAsync();
-
-            return orders;
-        }
-
-        public async Task<OrderDetail> GetOrderByIdAsync(Guid orderId)
-        {
-            var orderIdBytes = orderId.ToByteArray();
-
-            var order = await _orderContext.Order
-                .Where(x => _orderContext.Database.IsInMemory() ? x.Id.SequenceEqual(orderIdBytes) : x.Id == orderIdBytes)
-                .Select(x => new OrderDetail
-                {
-                    Id = new Guid(x.Id),
-                    ResellerId = new Guid(x.ResellerId),
-                    CustomerId = new Guid(x.CustomerId),
-                    StatusId = new Guid(x.StatusId),
-                    StatusName = x.Status.Name,
-                    CreatedDate = x.CreatedDate,
-                    TotalCost = x.Items.Sum(i => i.Quantity * i.Product.UnitCost).Value,
-                    TotalPrice = x.Items.Sum(i => i.Quantity * i.Product.UnitPrice).Value,
-                    Items = x.Items.Select(i => new Model.OrderItem
-                    {
-                        Id = new Guid(i.Id),
-                        OrderId = new Guid(i.OrderId),
-                        ServiceId = new Guid(i.ServiceId),
-                        ServiceName = i.Service.Name,
-                        ProductId = new Guid(i.ProductId),
-                        ProductName = i.Product.Name,
-                        UnitCost = i.Product.UnitCost,
-                        UnitPrice = i.Product.UnitPrice,
-                        TotalCost = i.Product.UnitCost * i.Quantity.Value,
-                        TotalPrice = i.Product.UnitPrice * i.Quantity.Value,
-                        Quantity = i.Quantity.Value
-                    })
-                }).SingleOrDefaultAsync();
-            
-            return order;
-        }
-
-        public async Task<IEnumerable<OrderSummary>> GetOrdersByStatusAsync(string statusName)
-        {
-            statusName = statusName?.Trim();
-
-            var query = _orderContext.Order
-                .Include(o => o.Status)
-                .Include(o => o.Items)
-                .ThenInclude(i => i.Product)
-                .AsQueryable();
-
-            if (!string.IsNullOrWhiteSpace(statusName))
-            {
-                query = query.Where(o => o.Status.Name == statusName);
-            }
-
-            var orders = await query
-                .Select(o => new OrderSummary
-                {
-                    Id = new Guid(o.Id),
-                    ResellerId = new Guid(o.ResellerId),
-                    CustomerId = new Guid(o.CustomerId),
-                    StatusId = new Guid(o.StatusId),
-                    StatusName = o.Status.Name,
-                    ItemCount = o.Items.Count,
-                    TotalCost = o.Items.Sum(i => i.Product.UnitCost * i.Quantity.Value),
-                    TotalPrice = o.Items.Sum(i => i.Product.UnitPrice * i.Quantity.Value),
-                    CreatedDate = o.CreatedDate
-                })
-                .ToListAsync()
-                .ConfigureAwait(false);
-
-            return orders;
-        }
-    }
-}
