+++ src/Order.Model/Order.Model.csproj
@@ -4,4 +4,9 @@
     <TargetFramework>net8.0</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Folder Include="DTOs\" />
+    <Folder Include="Requests\" />
+  </ItemGroup>
+
 </Project>
