+++ src/Order.Service/Interfaces/IOrderService.cs
@@ -1,0 +1,14 @@
+﻿using Order.Model;
+using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace Order.Service.Interfaces
+{
+    public interface IOrderService
+    {
+        Task<IEnumerable<OrderSummary>> GetOrdersAsync();
+        Task<OrderDetail> GetOrderByIdAsync(Guid orderId);
+        Task<IEnumerable<OrderSummary>> GetOrdersByStatusAsync(string status);
+    }
+}
